% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustAR.R
\name{AR}
\alias{AR}
\title{(Classical) AR Statistic}
\usage{
AR(Y, X, Z, W, beta0)
}
\arguments{
\item{Y}{A vector of dependent variables.}

\item{X}{A vector of endogenous covariates.}

\item{Z}{A matrix/vector of instrumental variables.}

\item{W}{A matrix/vector of control variables.}

\item{beta0}{A hypothesized beta0 value.}
}
\value{
A numeric scalar test statistic.
}
\description{
Computes the (classical) AR statistic introduced in Anderson and Rubin (1949).
}
