% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustAR.R
\name{huber_function_derivative}
\alias{huber_function_derivative}
\title{Derivative of the Huber function}
\usage{
huber_function_derivative(x)
}
\arguments{
\item{x}{A scalar}
}
\value{
A scalar
}
\description{
Derivative of the Huber function
}
