% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustAR.R
\name{id_function_derivative}
\alias{id_function_derivative}
\title{Derivative of Identity function}
\usage{
id_function_derivative(x)
}
\arguments{
\item{x}{A scalar}
}
\value{
A scalar
}
\description{
Derivative of Identity function
}
