% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustAR.R
\name{robustAR}
\alias{robustAR}
\title{Robust AR Statistic}
\usage{
robustAR(
  Y,
  X,
  Z,
  W,
  beta0,
  type = c("Huber", "Tukey", "OLS"),
  weighting = c("hat", "mcd", "no"),
  homoskedasticity = "no"
)
}
\arguments{
\item{Y}{A vector of dependent variables.}

\item{X}{A vector of endogenous covariates.}

\item{Z}{A matrix/vector of instrumental variables.}

\item{W}{A matrix/vector of control variables.}

\item{beta0}{A hypothesized beta0 value.}

\item{type}{Specifies which downweighting function to use: Huber, Tukey or OLS.}

\item{weighting}{Specifies which weighting method to use: mcd, hat matrix or no weights.}

\item{homoskedasticity}{Assume homoskedasticity: yes or no.}
}
\value{
A numeric scalar test statistic.
}
\description{
Computes the robust AR statistic introduced in Klooster and Zhelonkin (202x).
}
