% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustAR.R
\name{tukey_function_derivative}
\alias{tukey_function_derivative}
\title{Derivative of Tukey's biweight function}
\usage{
tukey_function_derivative(x)
}
\arguments{
\item{x}{A scalar}
}
\value{
A scalar
}
\description{
Derivative of Tukey's biweight function
}
